Employ pyproj 2 API for better use of PROJ 6
As part of the change is a safe-guard concerning transformations between
Danish and Greenlandic systems. WGS84 has been added as well as an
example of a CRS that is possible to transform to and from with both DK
and GL CRS's.

data.json has been updated to reflect proper axis ordering as governed
by PROJ and EPSG.