Add support for non-EPSG CRS definitions
Non-standard CRS's can now be used by leveraging a hub datum and custom
pipelines. The custom pipelines are build using the DK, and later on the
GL, ressource files from the proj-datumgrid-europe package. So far only
the CRS's DK:S34J, DK:S34S and DK:S45B has been added as a
demonstration. Adding more should not be complicated now that the
initial scaffolding is in place.

Closes #2