Refactor Transformation classes
This splits Transformation into three classes, one for the 2D, 3D and 4D
case. The main benefit of this is that we can document the parameters of
the three cases correctly. Before the 2D case would show up in the docs
with descriptions of all four coordinate components. Now the docs match
the domain of transformation.