Make WEBPROJ a installable Python package
This adds a setup.py script which installs the webproj package into
the Python userspace. As part of the setup a command line entry point id
added so that a development version of the webproj app can be started by
the command `webproj`.

The `webproj` command is neatly place in `app/main.py` which is the
default placement of a Flask application in the Docker image used for
this project.